Add dependence on kernel properties extension
The macro `SYCL_EXT_ONEAPI_FUNCTION_PROPERTY` is defined in this
extension.  In addition, it's not clear how viable free function
kernels are if we do not provide some way to decorate them with
things like the required sub-group size.  Therefore, it seems like
we may as well list the kernel properties extension as a dependency.