[SYCL][Graph] Skip prefetch/memadvise opencl tests
In https://github.com/intel/llvm/pull/11474 it was noted
in the documentation that the `prefetch` and `memadvise`
handler methods were not supported as graph nodes on the
OpenCL command-buffer backend.

However, we never disabled the E2E tests for this target,
which wasn't picked up by CI, as CI doesn't use the
command-buffer extension in its OpenCL config. Correct this
oversight.