[SYCL][Doc]Editorial changes "if_architecture_is" (#12258)
This commit makes some editorial changes to the "if_architecture_is"
extension specification. None of the APIs changed, so this has no effect
on the implementation. I'm making this in preparation for some new APIs
that I will soon add to this extension.

* Eliminate redundancy with the `architecture` enumeration. The code
synopsis no longer duplicates the enumerators. These are now entirely
defined in the table.

* Add some missing entries to the table defining the enumerators, and
fix some obvious cut-and-paste errors with the descriptions.

* Update the function descriptions to our new format, using subsections
`Constraints`, `Preconditions`, etc.

* Use our new style where we make a line break after every sentence.

---------

Co-authored-by: John Pennycook <john.pennycook@intel.com>