Add implementation note about diagnostics
Address the review comment about requiring an implementation to emit
a diagnostic if the kernel declaration violates the restrictions that
the spec mandates.  We're not sure yet if the spec should mandate a
diagnostic, but we do think that a high quality implementation should
do this.  Therefore, add an implementation note stating that we expect
DPC++ to issue a diagnostic in these cases.