Throw when device does not support USM allocation (#12446)
The USM allocation functions are changed to conform with the spec and
throw when certain usm allocation modes are not supported.
malloc_shared is left incomplete because it breaks a lot of tests and
its more reasonable to handle that in a later PR. In the meantime,
malloc_device and malloc_host have been changed to conform with the spec
and throw an exception when the corresponding USM aspect is not
supported. The test usm_pooling.cpp has also been updated to reflect
this query about the memory access properties in the L0 API so that it
knows to expect it in the output.