[SYCL] Fix resource leak related to SYCL_FALLBACK_ASSERT
https://github.com/intel/llvm/pull/6837 enabled asynchronous
buffer destruction for buffers constructed without host data. However,
initial fallback assert implementation in
https://github.com/intel/llvm/pull/3767 predates it and as such had to
place the buffer inside `queue_impl` to avoid unintended synchronization
point. I don't know if there was the same crash observed on the
end-to-end test added as part of this PR prior to
https://github.com/intel/llvm/pull/3767, but it doesn't even matter
because the "new" implementation is both simpler and doesn't result in a
crash.

I suspect that without it (with the buffer for fallback assert
implementation being a data member of `sycl::queue_impl`) we had a
cyclic dependency somewhere leading to resource leak and ultimately to
the assert in `DeviceGlobalUSMMem::~DeviceGlobalUSMMem()`.