Add func get_native_events as well as small fixes
Added an entry point for get_native_events, which gets the native events
that a HT is dependent on. Tests are passing. When using a parallel_for
and then a HT which depends on the PF, the SYCL RT still waits for the
PF native event to be entirely complete before dispatching the HT. This
is a problem with the model of synchronization in general with host
tasks. A new synchronization guarantee is needed in order to get around
this. Perhaps there is an opportunity to reinstate the interop task for
this reason. A solution would be if the interop task requires CGs that
it depends on to be enqueued before the HT is dispatched (so there are
native events). It would still require HT lambdas to be completed before
a new HT can be enqueued, so that add_native_events has populated the
SYCL RT with any events it might need.

Other changes:
- Move existing add_native_events to new dir. Test failing for accessor
  interop due to accessor interop being broken for CUDA atm. Refactor
  some stuff with a helper header.
- Change the void * to pi_native_handle for the interop_handle RT
  entry points.