[SYCL][COMPAT] Force device function to be inlined
Due to the way the inliner works, the launched function may become very large
and go above the inline threshold. This results with a short kernel which
only call one function.

The patch adds an always_inline on the call site to force the user function
to be inline in the SYCL kernel to reduce overhead.

Signed-off-by: Victor Lomuller <victor@codeplay.com>