Enable extra warnings, adjust code.
We would like to be able to build with -pedantic -Wcast-qual, the former
because it flags accidental use of non-standard language constructs, the
latter because it is also enabled by LLVM. These are now enabled, and
code is updated to address the warnings raised.

The C standard is bumped from C99 to C11 to account for the fact that we
were already using C11 language features, which now got flagged by
-pedantic. The move to C11 also enables the
checkDeprecatedOrUnsafeBufferHandling warning, which tells us not to use
scanf and to use scanf_s instead, despite scanf_s being optional and not
supported on platforms we support, and despite the use of scanf being
safe. Therefore, this warning is disabled.