[SYCL][Fusion] Handle fusion leading to synchronization issues  (#12538)
Do not allow fusion when one of the kernels has an explicit local size
and it requires ID remapping, i.e., it has a different number of
dimensions w.r.t. the fused ND-range or different global size in
dimensions [2, N). In this case, two work-items belonging to the same
work-group may not belong to the same work-group in the fused ND-range.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>

---------

Signed-off-by: Victor Perez <victor.perez@codeplay.com>