[SYCL] Make queue hash dependent on its unique ID
Using implementation pointer as the hash can lead to situations where
queues consistently get assigned the same hash as a queue that has been
destroyed (due to being allocated at the same address). This can cause
consistent hash collisions if a library uses hashes as a fast way
to detect a different queue where  the first one happens to be destoyed.