[SYCL] Make PropertySetRegistry owning it's content
Before this patch PropertySetRegistry used StringRefs
that point at external data that was supposed to outlive the instance of
PropertySetRegistry. It wasn't obvious and it might lead to memory bugs.
This patch makes PropertySetRegistry owning it's content as containers
usually do.

Also the previous implementation had a property that a content is
being printed in the order it was added. This patch removes this
property. However, the patch brings a sorted order in write() method
in order to make testing as easy as it was. Accordingly, all affected tests
were aligned to the sorted order.

Also documentation is aligned with doxygen BKSM.