[SYCL] Fix compiler crash. (#12324)
The compiler was crashing when the user requested fp-accuracy for the
functions in a call of the form f1(f2(f3 ...), where f1, f2 and f3 were
fpbuiltin but the innermost function didn't have an fpbuiltin. The
current builtinID was used instead of getting the builtinID from the
current function. that created a crash in the compiler.
This patch fixes the issue and renames the function
EmitFPBuiltinIndirectCall to MaybeEmitFPBuiltinofFD .