[syclcompat][CUDA] FIX UB in test / seq_cst requires sm_70 on CUDA (#12575)
Fix UB in test by using a single thread task.

A shared USM variable was being simultaneously written to by multiple
threads without using atomics. AFAIK this is generally not a well
defined program, and was leading to invalid values.

seq_cst also requires sm_70 on CUDA, so the compilation invocation is
updated to reflect this.
The CI device is >= sm_70 so it can use seq_cst. However this test did
not compile for >=sm_70.

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>