[SYCL][DOC] Add an overload for memory_required in joint_sorter (#11727)
From the implementation perspective it's important to have
`memory_required` in `default_sorters` dependent on the SYCL device. It
happens because `default_sorters` has no specified algorithm inside it
can depend on the specific SYCL device. `radix_sorters` has specific
algorithm (radix sorter) that doesn't depend on the backend.

Removed `constexpr` for default sorters since it's extra there. It's
still fine for radix sorters since the algorithm is fixed and can't
depend on the backend

Signed-off-by: Fedorov, Andrey <andrey.fedorov@intel.com>

---------

Signed-off-by: Fedorov, Andrey <andrey.fedorov@intel.com>