Rewrite tests that fail when usm_shared_allocations not supported. (#12636)
Nearly 40 E2E tests use malloc_shared and do not expect an exception to
be thrown in case of failure which is not what the spec says, namely,
malloc_shared should throw when the usm_shared_allocations aspect is not
supported by the device. At the moment malloc_shared simply returns
nullptr, however, once the implementation of malloc_shared is changed to
throw an exception these tests will produce failures.

This PR is the first of a series of PR's to rewrite these tests so that
they don't fail once malloc_shared is changed to conform to the spec.
These changes will only affect devices that do not support the relevant
aspect. This is done either:

- by having llvm-lit require that the aspect is supported for tests that
rely heavily on malloc_shared,
- by disabling, through if statements, sections of code that use
malloc_shared for tests that use it sparsely,
- by completely rewriting the malloc_shared code with equivalent
buffer/accessor code that is guaranteed to work on all platforms. The
latter approach is taken on tests that do not have the purpose of
testing malloc_shared but simply use it to allocate memory for the test.

These changes aim to eliminate test failures while maintaining test
coverage.
Subsequent PR's will omit this description and refer to this PR instead.