[SYCL][E2E] Fix the test script after changing sycl-ls output (#12672)
In PR#https://github.com/intel/llvm/pull/12596, we changed the output of
sycl-ls to use 'fpga' instead of 'acc'. This has caused several tests to
fail when fpga is available on the system. This PR fixes that.

In retrospect, we should have enabled testing on FPGA in pre-commit
testing as this bug could have been caught earlier.