Add force range rounding option and introduce new compiler flag
This commit adds a new preference for range rounding, force, such that if the
compile flag is used, only the range rounded parallel_for kernel will be generated.
This can make binaries smaller as there is no duplication of SYCL range kernels
across range rounded and unrounded versions.

I have also added the flag: -fsycl-range-rounding, which can have values: on, force
or disable. This flag aims to supercede the fsycl-disable-range-rounding flag.

I have also added to existing tests to check for the functionality of the new flag
and refactored the range rounding sycl-e2e test.