[SYCL] Creating version-agnostic copy of the SYCL import library
Ultimate goal is to get rid of most of the usages SYCL_MAJOR_VERSION in
clang, so that we don't need to keep that version in sync with SYCL RT.
Currently only one usage left which is related to windows-gnu triple which
is something we don't test (using compiler in mingw environment) and that
will need to be handled separately, as I wasn't able to succesfully
build/experiment with that.

Other reason why this is needed is because some users want to
explicitely link with sycl library and they don't want to update version
all the time.

It is possible to just generated version-agnostic .lib files but I
think we need to keep versioned files as well in case if somebody will
have a scenario when multiple releases are in the LIB env variable and
they want to link with specific version explicitely.