[SYCL] default context as default on windows (#12599)
Now that there have been improvements to Windows shutdown made by the
proxy loader, host task handling, and the unified runtime, we are
finally able to have Windows support a "default" context. This makes any
routine that loops over queue creation, etc perform much faster.

Note that on Windows the UR L0 Leak check does not work correctly with
the default contexts on Windows. This is because on Windows the release
of the plugin DLLs races against the release of static global variables
(like the default context). This has been reported.

The tests that use this feature have been modified to use a new
`%{l0_leak_check}` substitution that will NOT use default context on
Windows.