[SYCL][Bindless][Exp] 3D images accept 3 component vecs instead of 4 (#12581)
Update read/write image functions to only accept coords with three
arguments instead of the current four, for 3D images.

Before this patch, when reading or writing to 3D bindless images, 4D
coordinates had to be used, where the last coord is always ignored. This
patch aligns bindless images to the rest of SYCL and other solutions.