[ESIMD] Fix atomic_update() implementation for N=16 and N=32 on Gen12
atomic_update() for USM and ACC were lowered to SVM/DWORD atomic
intrinsics even though the HW instructions on Gen12 supported only
N up to 8 for USM and up to 16 for ACC.

GPU had legalization pass for N that split longer vectors to smaller and available in HW.
That GPU optimization/legalization workes incorrectly for USM as it
splits longer vectors assuming instruction is available for N=16 in case
of USM, which is not correct.

The patch here implements splitting of N=16 and N=32 cases for
atomic_update(usm, ...) to N=8 vectors until GPU fixes the legalization
for USM atomic_update.

Signed-off-by: Klochkov, Vyacheslav N <vyacheslav.n.klochkov@intel.com>