[SYCL] Fix malloc shared by throwing when usm_shared_allocations not supported (#12700)
Final PR in the series of #12636.
Refer to it for a description.
After a discussion with @AlexeySachkov we've decided its best to not
rewrite USM and syclcompat tests with buffers/accessors. For USM, the
reason is obvious and for syclcompat you can reach out to Alexey.
Therefore, these tests are handled using if statements or requring
aspect to be supported.
Once this PR is merged, the behavior of malloc_shared will be to throw
if the usm_shared_allocations is not supported which is conformant with
the spec.