[SYCL] Add a CUDA compatibilty mode
This patch enables CUDA mode at the same time as the SYCL mode.
This allows the compiler to define CUDA macros and add implicit defines.

To enable the mode the user adds -fsycl-cuda-compat to the command line.
By default this mode is set to off. The flag is only used for the NVPTX backend.

The intent is to help to transition from CUDA to SYCL. Using this mode enable
a SYCL application to reuse CUDA functionnalities, especially fast paths that
are guarded by __CUDA_ARCH__.

Signed-off-by: Victor Lomuller <victor@codeplay.com>