[SYCL] Fix ballot_group when the sub-group is not full size
Not all sub-groups are necessarily the max size of sub-groups in the
kernel invocation. As such, non-uniform groups should handle these
sub-groups properly. However, due to how the mask for the false-group
in ballot_group creates its mask, it thinks it has full 32-element size
no matter how big the actual sub-group is. This commit fixes this issue.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>