Refactor device initialization
Some logic for device initialization was split across platform
init and device init. This duplicated some calls to get info
for max block dim in Y and Z dimension. This rectifies that
and makes sure the max block dims are only queried once, which
is called from the device constructor.