[EXP][CMDBUF] Move event reset commands to dedicated cmd-list
Create a new command-list to reset all the events of the enqueued graph cmd. This allows us to move the enqueueing of reset commands from the command-buffer enqueue function to the finalize function, and allow us to perform this task only once.
Move signal event reset from main cmd-list to reset cmd-list.
Move the documentation to DPC++ design doc.