[SYCL] Make queue hash dependent on its unique ID (#12578)
Using implementation pointer as the hash can lead to situations where
queues consistently get assigned the same hash as a queue that has been
destroyed (due to being allocated at the same address). In some cases,
this can cause consistent hash collisions if a library uses these hashes
as a fast way to detect a different queue being passed to it.