[SYCL] Make 'supportAcc' a template variable in getSyclDeviceTypeMap (#12775)
The output of getSyclDeviceTypeMap() is different for
SYCL_DEVICE_ALLOWLIST and ONEAPI_DEVICE_SELECTOR. However, currently, we
return a reference to a 'static const' type, which causes calls to
getSyclDeviceTypeMap() return the same value, irrespective of
'supportAcc'. This results in several test failures.

This PR makes 'supportAcc' a template parameter instead.