[CI] Run SYCL-CTS with the latest nightly
Add a new workflow to run SYCL-CTS when new nightly build is available.
This should help us to handle new failures in advance, monitor new
suites and so on.

To start with running only reduced scope of tests (cts_exclude_filter is
big enough) on unused "debug" runner (linux+cpu).