[SYCL][Graph] Skip Graph tests based on sycl-ls output
The graph extension tests are currently skipped during execution
for devices which don't support the graphs extension. However,
this early return causes the tests to be reported as passed and
makes it hard from looking at the results to know if the tests
actually stressed the graphs code or not.

Improved this situation by modifying `sycl-ls --verbose` to output
the `sycl_ext_oneapi_graph` string
when a device supports the graph extension. We can then use this to
inform the LIT config and set a requirement for the `graph-extension`
feature in the tests, similar to what is done for aspects. This enables the
tests to be obviously skipped for devices that don't support graphs.

To enable setting this requirement in `lit.local.cfg` files some extra
directories have been created, in particular `UnsupportedDevice` which
doesn't have a requirement on `graph-extension` as the tests verify
expected behaviour with unsupported devices.