Struct as result type in OpGroupNonUniformShuffleDown translation fix (#2339)
The contract between some frontends and Translator is that
calls to functions which are matching to hardcoded name are
converted to specific Ops in SPIR-V. The translation from LLVM to SPIR-V of
OpGroupNonUnfiromShuffleDown call was done incorrectly. The struct member
which is correct target for this op was left unwrapped which resulted in
SPIR-V nonconforming with specification. The result of this were later problems
with e.g. SPIR-V to LLVM translation.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/6b3ec4166a63039