[SYCL][E2E] Fix USM/usm_leak_check.cpp test for multi-card PVC (#12817)
**Problem:**
USM/usm_leak_check.cpp fails on multi-card PVC because the number of
zeMemAlloc* and zeMemFree calls doesn't match what the test expects.
This is mainly because of the recent changes to USM memory allocation,
like a different memory pool logic.

**Proposed Fix**
Since the objective of this test is to just check for USM leaks, I don't
think that we even have to check the output of UR_L0_LEAKS_DEBUG=1. It
is my understanding that UR_L0_LEAKS_DEBUG=1 will abort upon finding a
mismatch in the number of alloc/free calls. So, this PR removes explicit
checks for zeMemAlloc* and zeMemFree calls.