[Driver][SYCL] Improve support for -fsycl-link with AOT
Support for -fsycl-link was restricted to only providing device linking
and providing the final device binary for JIT.  These changes expand on
that functionality to pull in the ability to also perform offline
compilation steps to support AOT targets.

Example:

  clang++ -fsycl -fsycl-targets=spir64_gen -c file.cpp
  clang++ -fsycl -fsycl-link -fsycl-targets=spir64_gen
          -Xsycl-target-backend "-device skl" file.o -o newfile.o
  clang++ file.o newfile.o -o finalexe

'newfile.o' here is a final device binary for GPU that is wrapped and
allowed to be linked in as a regular host object.

The changes here are significant, performing adjustments on when the
device linking occurs within the compilation toolchain within the
driver.  This improves the general dependency flow, allowing for the
device linking to occur in a common location for -fsycl-link and normal
link steps.  Due to this, a number of tests relying on phase checking
required to be tweaked.