[test] Fix infinite loop when we run out of input
d6930c70bd7 introduced Status.INPUT_END but didn't introduce the means
to handle that eventuality. It looks like the intention might have been
to exit with success if we've got to the end of input.
However, if we run out of input and still have matches pending, I'm
treating this as a failure