[ESIMD][NFC] Rename L3 cache hint parameters to L2 to avoid confuses
The experimental memory API used L3 to name the cache that always was L2
due to some historical reasons.
The new non-experimental API names the same cache level as L2.

This patch just renames the old L3 params to L2 to unify the naming in
experimental and non-experimental namespaces and avoid the mess in
cache names.

Signed-off-by: Klochkov, Vyacheslav N <vyacheslav.n.klochkov@intel.com>