[SYCL] Fix component device as descendent device handling
According to `sycl_ext_oneapi_composite_device` documentation a
component device is considered to be a descendent from a composite
device it belongs to.

This implies that context created for a composite device work with
component devices as well.

In order to achieve this, this patch changes `context_impl` constructor
to include all component devices into context's device list if a
composite device was passed by user to the constructor.