Clarify iteration space and builtin vars in SPIR-V
Also disallows launching a SPIR-V kernel as a simple "range" kernel.
This seems consistent with our view that a "range" kernel is not just a
degenerate form of an nd-range kernel.  Since SPIR-V kernels always
have access to nd-range features, it does not make sense to launch them
as range kernels.  This is also consistent with our decision to limit
SYCL free function kernels to "nd-range" and "single-task" forms.