[FPGA][SYCL] Add support for memory attributes on device_global variables (#12785)
Currently we allow FPGA memory attributes on const variables. This patch
relaxes the restriction that permits the attributes when applied on
non-const device_global variables and updates the error message to
properly convey which types of variables are allowed to have attributes.
Also this patch relaxes the restriction that permits the private_copies
attribute when applied on const variables. Previously we had restriction
that the private_copies attribute would work with local non-const
variables only.

---------

Signed-off-by: Soumi Manna <soumi.manna@intel.com>