[CMake] Support external adapter source dirs
Allow the configuration of external source directory for any of the
existing adapters. This is done via a CMake cache variable which if set
will is used as the `source_dir` argument to `add_subdirectory()`, the
`binary_dir` argument will also be set to the same location as a normal
in-tree build. This is done via the addition of the new `add_adapter()`
helper function which also hoists the logic to enable/disable each
adapter.