[UR] Support fetch adapter source individually
This patch and its counterpart in oneapi-src/unified-runtime#1410 add
CMake support for fetching an individual Unified Runtime adapter's
source code from a different repo/tag combination using the new
`fetch_adapter_source()` CMake function. Only the source is cloned, it
is not added to the build directly.

Instead, the path to the adapter source is passed into Unified Runtime
clone described by the `UNIFIED_RUNTIME_REPO` and `UNIFIED_RUNTIME_TAG`
CMake variables. This clone is the source of truth for the Unified
Runtime API and drives the build of the external adapter source.

Using `fetch_adapter_source()` is optional.