[CI] Fix E2E workflows on PR trigger
By default 'issue_comment' trigger gets the context of a default branch.
If we want to test the actual PR, on which we created/edited a comment
we have to fetch and checkout a special ref for PR's merge commit.

Note, that 'pull/<no>/merge' ref is available as long as the PR
is not merged. This commit do not take this under consideration, since
this case should be fairly uncommon and would obfuscate the workflow.