[SYCL][Docs] Invert group numbering for ballot groups (#12905)
The current extension for non-uniform groups state that ballot groups
are numbered 0 if the predicate was true and 1 if the predicate was
false. However, the implementation has the inverse numbering. This
commit changes the extension to make the numbering consistent with the
implementation, in turn making the group numbering directly convertible
to a boolean value equal to the corresponding predicate.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>