[SYCL][Doc] Clarify WI funcs in kernel compiler (#12891)
Clarify the SPIR-V and OpenCL kernel compiler specifications to explain
how the SYCL iteration space maps to the SPIR-V / OpenCL C languages
and clarify that these kernels can use their normal language mechanisms
to find the current work-item's position in the iteration space.

We also disallow launching a SPIR-V or OpenCL C kernel as a simple
"range" kernel.  This seems consistent with our view that a "range"
kernel is not just a degenerate form of an nd-range kernel.  Since
SPIR-V and OpenCL C kernels always have access to nd-range features, it
does not make sense to launch them as range kernels.  This is also
consistent with our decision to limit SYCL free function kernels to
"nd-range" and "single-task" forms.

Some other cleanup of these specifications also:

* Clarify what happens when a `local_accessor` is passed as a kernel
  argument to a SPIR-V or OpenCL kernel.  This was causing some
  confusion from users.

* Reformat the table in the OpenCL spec describing kernel arguments
  so that it has the same layout as the equivalent SPIR-V table.

* Fix the name of the OpenCL header file in the example.