Introduce a utility to walk polymorphic linked lists
Add `find_stype_node` for walking polymorphic linked lists looking for a
particular type.

The implementation here works with an auto-generated compile time map,
that links a given type to the structure enumeration tag. The
implementation simply walks the list looking for the .stype implied by
the template parameter type and casts it to the expected type, then
returning it.

It's one of those unfortunate cases where you can write pretty nasty
implementation code that makes the user code much much nicer. In this
case the user doesn't need to worry about the `.types` at all, and the
const-void casts can be eliminated from user code.