[SYCL][SCLA] Add basic `sycl_ext_oneapi_private_alloca` functionality
After a1050551296b implementing CodeGen capabilities for
`sycl_ext_oneapi_private_alloca`, this patch handles the generated
intrinsic in `sycl-post-link` for targets with native specialization
constants support.

Headers for the new extension are also added, as well as a feature
test macro.

`multi_ptr` definitions in the SYCL headers are annotated with the
`__sycl_detail__::sycl_type` to be detected by the frontend.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>