[SYCL] Clean up sub-group class API (#12945)
This PR cleans up the API of the sub-group class by either removing
functions that do not appear in the spec or marking them deprecated. The
only exception is the set of load/store functions, which after some
discussion with the spec team, it was suggested to leave them as is for
now.