[SYCL][NATIVECPU] Avoid overwriting offload kind on Native CPU (#12971)
For Native CPU, the host and the device triples are the same. This means
that the cache entries for `BuildJobsForAction` are distinguished only
by the offload kind (`-host` or `-sycl`). Since the offload kind is
being overwritten when processing `OffloadUnbundlingJobAction`, this
lead to cache entries being overwritten on Native CPU, with
corresponding build failures.