[CL] Atomic fence scope capability for Intel FPGA driver
This patch implements a workaround for the Intel FPGA driver which is
currently an OpenCL 1.2 device which also supports
`UR_MEMORY_SCOPE_CAPABILITY_FLAG_DEVICE` rather than the baseline for
OpenCL 1.2 devices.

The workaround unconditionally queries for
`CL_DEVICE_ATOMIC_FENCE_CAPABILITIES`, if it is successful the value is
used. Otherwise the error code is ignored and the baseline capabilities
are returned.